Fix resolution of components by complex selector
Attribute selectors containing quotes were throwing an error when trying
to get the `[data-component="${nameOrSelector}"]` elements as they were
multiple quotes in the selector, i.e. `[data-component="a[href^="#"]"]`.

This commit adds a simple try/catch combo to avoid throwing an error as
there is a fallback to the component's resolution.