Add support for services with event handlers only
Before, if the service method was not defined, the service was not
enabled. With this commit, it is now possible to use services only via
their events, without defining the method:

```js
class Component extends Base {
	static config = {
		name: 'Component',
	};

	mounted() {
		this.$on('ticked', () => {});
	}
}
```

This feature is meant to be used by decorators, it is recommended to use
the methods API to use services in a component.