Use damping on the current value instead of the progress
As the progress value is clamped to a 0–1 interval, the damp factor and
precisions value can be off when working with large elements.

For example, for an element of 100 pixels, progress from 0 to 0.1
represents 10px, but for an element of 1000 pixels, it represents 100px.
To keep the same precision between the two elements, we would need to
adjust the `dampPrecision` property. By applying the damping on the
sizes in pixels directly, we get rid of this problem, as the base unit
is fixed in pixels.