Fix $mount and $update method never resolving
Async components were added to the list of component that should be
awaited before mounting or updating the parent, creating some long
hanging promises that might never resolve. This commit switch back to
the previous behavior of not waiting for async components to be resolved
before considering all children mounted (which is the point of async
components).

Fix: #542