Refactor reference to components' instances on the root element
Using the class of a component as a key for the storage of its instance
is error prone when working with inheritance, as a component with the
same name can be mounted twice on a single element.

Switching to using the name of the component will help stabilize this
feature and is aligned with the latest evolution towards name based
query for components.